+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Added options for specifying the Slack access token, channel name and port to run the server on. ([#6](https://github.com/hjwylde/werewolf/issues/6))
 
+*Minor*
+
+* Added a `--debug` flag. ([#4](https://github.com/hjwylde/werewolf/issues/4))
+
 ### v0.1.0.0
 
 *Major*