+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 * Added a `--validation-token` option and validation. ([#9](https://github.com/hjwylde/werewolf/issues/9))
 * Added a `--debug` flag. ([#4](https://github.com/hjwylde/werewolf/issues/4))
 
+*Revisions*
+
+* Fingers crossed fixed a bug with UTF-8 parsing of an accented e. ([#10](https://github.com/hjwylde/werewolf/issues/10))
+
 ### v0.1.0.0
 
 *Major*